move center out variants to another file
