+++ built_in_tasks/target_capture_task_xt.py
@@ -4,10 +4,14 @@ Target capture tasks with additional features
 '''
 import numpy as np
 import random
+import os
 
-from riglib.experiment import traits
 from .target_graphics import *
 from .target_capture_task import ScreenTargetCapture
+from riglib.experiment import traits
+from riglib.audio import AudioPlayer
+
+audio_path = os.path.join(os.path.dirname(__file__), '../riglib/audio')        
 
 class ScreenReachAngle(ScreenTargetCapture):
     '''
@@ -379,6 +383,7 @@ class SequenceCapture(ScreenTargetCapture):
             yield indices, targs
 
 
+
 class HandConstrainedEyeCapture(ScreenTargetCapture):
     '''
     Saccade task with holding another target with hand. Subjects need to hold an initial target with their hand. 
@@ -686,6 +691,7 @@ class HandConstrainedEyeCapture(ScreenTargetCapture):
             yield indices, targs
 
 
+
 class ScreenTargetCapture_Saccade(ScreenTargetCapture):
     '''
     Center-out saccade task. The controller for the cursor position is eye position.
@@ -719,3 +725,153 @@ class ScreenTargetCapture_Saccade(ScreenTargetCapture):
     def _start_hold(self):
         super()._start_hold()
         self.targets[self.target_index].sphere.color = target_colors[self.fixation_target_color] # change target color in fixating the target
+
+
+
+class ScreenTargetCapture_ReadySet(ScreenTargetCapture):
+
+    '''
+    Center out task with ready set go auditory cues. Cues separated by 500 ms and participant is expected to move on final go cue. Additionally, participant must move out
+    of center circle (mustmv_time) parameter or there will be an error. 
+    '''
+    
+    status = dict(
+        wait = dict(start_trial="target"),
+        target = dict(enter_target="hold", timeout="timeout_penalty"),
+        hold = dict(leave_target="hold_penalty", hold_complete_center="prepbuff", hold_complete_periph='reward'),
+        prepbuff = dict(leave_target="hold_penalty", prepbuff_complete="delay"),
+        delay = dict(leave_target="delay_penalty", delay_complete="leave_center"),
+        leave_center = dict(leave_target="targ_transition", mustmv_complete="tooslow_penalty"),
+        targ_transition = dict(trial_complete="reward", trial_abort="wait", trial_incomplete="target"),
+        timeout_penalty = dict(timeout_penalty_end="targ_transition", end_state=True),
+        hold_penalty = dict(hold_penalty_end="targ_transition", end_state=True),
+        tooslow_penalty = dict(tooslow_penalty_end="targ_transition", end_state=True),
+        delay_penalty = dict(delay_penalty_end="targ_transition", end_state=True),
+        reward = dict(reward_end="wait", stoppable=False, end_state=True)
+    )
+
+    #the sum of the prepbuff & the delay time should be equal to the length of the ready_set_sound file 
+    # the delay time corresponds to the amount of time the peripheral target is displayed, and the prepbuff time then makes up the difference 
+    wait_time = traits.Float(1., desc="Length of time in wait state (inter-trial interval)")
+    prepbuff_time = traits.Float(.2, desc="How long after completing center target hold before peripheral target appears")
+    mustmv_time = traits.Float(.2, desc="Must leave center target within this time after auditory go cue.")
+    tooslow_penalty_time = traits.Float(1, desc="Length of penalty time for too slow error")
+    files = [f for f in os.listdir(audio_path) if '.wav' in f]
+    ready_set_sound = traits.OptionsList(files, desc="File in riglib/audio to play on each trial for the go cue")
+    tooslow_penalty_sound = traits.OptionsList(files, desc="File in riglib/audio to play on each must move penalty") #hold penalty is normally incorrect.wav
+    shadow_periph_radius = traits.Float(0.5, desc = 'additional radius for peripheral target')
+    periph_hold = traits.Float(0.2, desc = "Hold time for peripheral target")
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.ready_set_player = AudioPlayer(self.ready_set_sound)
+        self.tooslow_penalty_player = AudioPlayer(self.tooslow_penalty_sound)
+        self.pseudo_reward = 0
+        
+    ###Test Functions ###
+
+    def _test_start_trial(self, time_in_state):
+        '''Start next trial automatically. You may want this to instead be
+            - a random delay
+            - require some initiation action
+        '''
+        return time_in_state > self.wait_time
+    
+    def _test_hold_complete_center(self, time_in_state):
+        '''
+        Test whether the center target is held long enough to declare the
+        trial a success 
+
+        Possible options
+            - Target held for the minimum requred time (implemented here)
+            - Sensorized object moved by a certain amount
+            - Sensorized object moved to the required location
+            - Manually triggered by experimenter
+        '''
+        return self.target_index == 0 and time_in_state > self.hold_time
+    
+    def _test_hold_complete_periph(self, time_in_state):
+        '''
+        Test whether the peripheral target is held long enough to declare the
+        trial a success 
+
+        Possible options
+            - Target held for the minimum requred time (implemented here)
+            - Sensorized object moved by a certain amount
+            - Sensorized object moved to the required location
+            - Manually triggered by experimenter
+        '''
+        return self.target_index == 1 and time_in_state > self.periph_hold
+    
+    def _test_prepbuff_complete(self, time_in_state):
+        '''
+        Test whether the center target is held long enough to transition from the prepbuff time to the delay state. 
+        The delay state will display the peripheral target so this state just requires a center hold.  
+
+        '''
+        return time_in_state > self.prepbuff_time
+    
+    def _test_mustmv_complete(self, time_in_state):
+        '''
+        Test whether the target is exited in time. Return of true for mustmv sends to penalty state.  
+
+        Possible options
+            - Target left before the minimum requred time (implemented here)
+            - Sensorized object moved by a certain amount
+            - Sensorized object moved to the required location
+            - Manually triggered by experimenter
+        '''
+        return time_in_state > self.mustmv_time
+    
+    def _test_tooslow_penalty_end(self, time_in_state):
+        return time_in_state > self.tooslow_penalty_time
+    
+    def update_report_stats(self): #add holds completed metric to report stats
+        super().update_report_stats()
+        self.reportstats['Audio Completed'] = self.calc_state_occurrences('leave_center') #count if delay state completed
+        self.reportstats['Pseudo Reward'] = self.pseudo_reward + self.reward_count
+
+    ### State Functions ###
+    def _start_prepbuff(self):
+        self.sync_event('CUE') #integer code 112
+        self.ready_set_player.play() 
+
+    def _start_leave_center(self):
+        self.sync_event('CURSOR_LEAVE_TARGET') #integer code 96
+    
+        
+    def _start_hold_penalty(self):
+        self.pseudo_success() #run before increment trials to prevent reseting of trial index 
+        if hasattr(super(), '_start_hold_penalty'):
+            super()._start_hold_penalty()
+        self.ready_set_player.stop()
+    
+    def _start_delay_penalty(self):
+        if hasattr(super(), '_start_delay_penalty'):
+            super()._start_delay_penalty()
+        self.ready_set_player.stop()
+    
+    def _start_timeout_penalty(self):
+        self.pseudo_success() #run before increment trials to prevent reseting of trial index
+        super()._start_timeout_penalty()
+
+    def _start_tooslow_penalty(self):
+        self._increment_tries()
+        self.sync_event('OTHER_PENALTY') #integer code 79
+        self.tooslow_penalty_player.play()
+        self.ready_set_player.stop()
+        self.jack_count = 0
+        # # Hide targets
+        for target in self.targets:
+            target.hide()
+            target.reset()
+    
+    def _end_tooslow_penalty(self):
+        self.sync_event('TRIAL_END')
+    
+    def pseudo_success(self): #function to measure almost success
+        if self.target_index == 1: #if peripheral target is displayed 
+            target_buffer_dist = self.target_radius + self.shadow_periph_radius - self.cursor_radius #combined radius 
+            dist_from_targ = np.linalg.norm(self.plant.get_endpoint_pos() - self.targs[self.target_index]) #vector difference
+            if dist_from_targ <= target_buffer_dist:
+                self.pseudo_reward += 1 #increment if cursor position is less than the shadow radius plus radius
